+++ packages/geostreaming/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [Github-59](https://github.com/geostreams/geodashboard/issues/59)
 
 ## 3.8.0 - 2021-07-13
+## Added
+- New configuration options added for visualizations (forceVega, defaultStartAtZero, defaultSameTimeScale)
+- Vega Visualizations added (Multi-Line Chart, Line Chart, Box-Whisker Plot)
 
 ### Changed
 - Created Search page with filters to replace old search/download page